+++ src/markets/germany/rewe/rewe_scraper_citylist.py
@@ -1,0 +1,8 @@
+citylist = [
+    ["Kaufbeuren", ["Kaufbeuren", "Kaufbeuren-Neugablonz"]],
+    ["Pforzen", ["Pforzen"]],
+    ["Füssen", ["Füssen"]],
+    ["Memmingen", ["Memmingen"]],
+    ["Kempten", ["Kempten", "Kempten (Allgäu)"]],
+    ["Kempten St. Mang", ["Kempten St. Mang", "Kempten-St. Mang", "Kempten (Allgäu)-St. Mang"]]
+]